Ignore Foresporsel tag in XML of type dialog_notat/henvendelse fra lege
This is necessary because EPJs were allowed to send incorrect data to
previous system EIA. Even if the EPJ did send correct XML we're currently
not using the information.

In the future, when we start threading messages, we'll need to make sure
they also send the required information to identify the original message
they're responding to.