Look for behandler name in Notat if not in MsgHead.
If it isn't there either, the name will be empty. The signing lege will
always be named in the PDF, and the TSS_ID, if present, will also help
identify who sent the message in Arena.

The behandler should always be in the MsgHead, but some chiropractor EPJs
have been allowed to send it in Notat in Eia, so we need to support it.