Replace kuhr sar with smtss and smgcp. (#185)
The side effect of updating emottak subscription is now handled by calling
smgcp.

Get tss from emottak endpoint. The alternative infotrygd endpoint is not
relevant because it might return incorrect info. The infotrygd tss ids that
have a avd01 value on them are entries in TSS from imports from HPR,
and do not have a valid address or legekontor associated with them. In the
event that there is no match from the emottak endpoint, it might be useful
to use these TSS IDs, which is the case for sykmeldinger who needs a tss to
store in infotrygd, but this is not necessary for dialogmeldinger or for
updating emottak subscription. In these cases a correct address is
more important.

Co-authored-by: June Henriksen <june.henriksen2@nav.no>